Add Users Authentication

+++ app/helpers/ideas_helper.rb
@@ -1,7 +1,7 @@
 module IdeasHelper
   def check_access(idea)
     out = ''
-    if user_signed_in?
+    if user_signed_in? && current_user.name == idea.user.name
       out << link_to('Show', idea)
       out << link_to('Edit', edit_idea_path(idea))
       out << link_to('Destroy', idea, method: :delete, data: { confirm: 'Are you sure?' })
